requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
duckduckgo-search
langchain
langchain-community
faiss-cpu
groq
pinecone-client
requests